+++ utils/ModuleUtils.js
@@ -1,0 +1,5 @@
+export default class ModuleUtils {
+    static get version() {
+        return JSON.parse(FileLib.read("RiccioFishingUtils", "metadata.json")).version
+    }
+}